fix PySide2 tests
